requirements.txt
Flask==3.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
psycopg2-binary==2.9.10
Pillow==11.1.0
opencv-python-headless==4.11.0.86
numpy==2.2.3
invisible-watermark==0.2.0
python-dotenv==1.0.1
eventlet==0.37.0
